chore(dapp): Refactor move IotaNamesClientContext under src/contexts
