+++ dapp/src/contexts/index.ts
@@ -1,1 +1,5 @@
+// Copyright (c) 2025 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+export * from './IotaNamesClientContext';
 export * from './ThemeContext';
