--- dapp/src/providers/contexts/index.ts
@@ -1,4 +1,0 @@
-// Copyright (c) 2025 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-export * from './IotaNamesClientContext';
